fix: jq error for non existing files

+++ tools/apigee-sackmesser/cmd/report/helpers/apiproducts.sh
@@ -35,23 +35,24 @@ echo "<th data-sortable=\"true\" data-field=\"enabled\">Aproval Type</th>" >> "$
 echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/api-products".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/api-products".json | while read i; do 
+        name=$(echo "$i" | jq -r '.name')
+        envs=$(echo "$i" | jq -r '.environments[]')
+        if [ "$opdk" == "T" ]; then
+            proxies=$(echo "$i" | jq -r '.proxies[]')
+        elif [ "$apiversion" = "google" ]; then
+            proxies=$(echo "$i" | jq -r '.operationGroup.operationConfigs[]' | jq -r '[.apiSource, .operations[].resource] | join(": ")')
+        fi
+        approvalType=$(echo "$i" | jq -r '.approvalType')
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/api-products".json | while read i; do 
-    name=$(echo "$i" | jq -r '.name')
-    envs=$(echo "$i" | jq -r '.environments[]')
-    if [ "$opdk" == "T" ]; then
-        proxies=$(echo "$i" | jq -r '.proxies[]')
-    elif [ "$apiversion" = "google" ]; then
-        proxies=$(echo "$i" | jq -r '.operationGroup.operationConfigs[]' | jq -r '[.apiSource, .operations[].resource] | join(": ")')
-    fi
-    approvalType=$(echo "$i" | jq -r '.approvalType')
-
-    echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-    echo "<td>$name</td>"  >> "$report_html"
-    echo "<td>$envs</td>"  >> "$report_html"
-    echo "<td>$proxies</td>" >> "$report_html"
-    echo "<td>$approvalType</td>" >> "$report_html"
-    echo "</tr>"  >> "$report_html"
-done
+        echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+        echo "<td>$name</td>"  >> "$report_html"
+        echo "<td>$envs</td>"  >> "$report_html"
+        echo "<td>$proxies</td>" >> "$report_html"
+        echo "<td>$approvalType</td>" >> "$report_html"
+        echo "</tr>"  >> "$report_html"
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
