+++ tools/apigee-sackmesser/cmd/report/helpers/flowhooks.sh
@@ -35,30 +35,32 @@ echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/flowhooks".json | while read i; do 
-    if [ "$opdk" == "T" ]; then
-        name=$(echo "$i" | jq -r '.name')
-    elif [ "$apiversion" = "google" ]; then
-        name=$(echo "$i" | jq -r '.flowHookPoint')
-    fi
-    
-    sharedFlow=$(echo "$i" | jq -r '.sharedFlow')
-    _continueOnError=$(echo "$i" | jq -r '.continueOnError')
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/flowhooks".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/flowhooks".json | while read i; do 
+        if [ "$opdk" == "T" ]; then
+            name=$(echo "$i" | jq -r '.name')
+        elif [ "$apiversion" = "google" ]; then
+            name=$(echo "$i" | jq -r '.flowHookPoint')
+        fi
+        
+        sharedFlow=$(echo "$i" | jq -r '.sharedFlow')
+        _continueOnError=$(echo "$i" | jq -r '.continueOnError')
 
-    if [ $_continueOnError = true ]
-        then
-            continueOnError="✅"
-        else
-            continueOnError="❌"
-    fi
-    if [ $name != null ]
-        then
-            echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-            echo "<td>$name</td>"  >> "$report_html"
-            echo "<td>$sharedFlow</td>" >> "$report_html"
-            echo "<td>$continueOnError</td>" >> "$report_html"
-            echo "</tr>"  >> "$report_html"
-    fi            
-done
+        if [ $_continueOnError = true ]
+            then
+                continueOnError="✅"
+            else
+                continueOnError="❌"
+        fi
+        if [ $name != null ]
+            then
+                echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+                echo "<td>$name</td>"  >> "$report_html"
+                echo "<td>$sharedFlow</td>" >> "$report_html"
+                echo "<td>$continueOnError</td>" >> "$report_html"
+                echo "</tr>"  >> "$report_html"
+        fi            
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
