+++ tools/apigee-sackmesser/cmd/report/helpers/keystores.sh
@@ -36,18 +36,20 @@ echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/keystores".json | while read i; do 
-    name=$(echo "$i" | jq -r '.name')
-    aliasCount=$(echo "$i" | jq -r '.aliases | length')
-    keyCount=$(echo "$i" | jq -r '.keys | length')
-    certCount=$(echo "$i" | jq -r '.certs | length')
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/keystores".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/keystores".json | while read i; do 
+        name=$(echo "$i" | jq -r '.name')
+        aliasCount=$(echo "$i" | jq -r '.aliases | length')
+        keyCount=$(echo "$i" | jq -r '.keys | length')
+        certCount=$(echo "$i" | jq -r '.certs | length')
 
-    echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-    echo "<td>$name</td>"  >> "$report_html"
-    echo "<td>$keyCount</td>" >> "$report_html"
-    echo "<td>$aliasCount</td>" >> "$report_html"
-    echo "<td>$certCount</td>" >> "$report_html"
-    echo "</tr>"  >> "$report_html"
-done
+        echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+        echo "<td>$name</td>"  >> "$report_html"
+        echo "<td>$keyCount</td>" >> "$report_html"
+        echo "<td>$aliasCount</td>" >> "$report_html"
+        echo "<td>$certCount</td>" >> "$report_html"
+        echo "</tr>"  >> "$report_html"
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
