+++ tools/apigee-sackmesser/cmd/report/helpers/keyvaluemaps.sh
@@ -36,23 +36,25 @@ echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/kvms".json | while read i; do 
-    name=$(echo "$i" | jq -r '.name')
-    _encrypted=$(echo "$i" | jq -r '.encrypted')
-    keyCount=$(echo "$i" | jq -r '.entry | length')
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/kvms".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/kvms".json | while read i; do 
+        name=$(echo "$i" | jq -r '.name')
+        _encrypted=$(echo "$i" | jq -r '.encrypted')
+        keyCount=$(echo "$i" | jq -r '.entry | length')
 
-    if [ $_encrypted = true ]
-        then
-            encrypted="✅"
-        else
-            encrypted="❌"
-    fi
+        if [ $_encrypted = true ]
+            then
+                encrypted="✅"
+            else
+                encrypted="❌"
+        fi
 
-    echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-    echo "<td>$name</td>"  >> "$report_html"
-    echo "<td>"$encrypted"</td>"  >> "$report_html"
-    echo "<td>$keyCount</td>" >> "$report_html"
-    echo "</tr>"  >> "$report_html"
-done
+        echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+        echo "<td>$name</td>"  >> "$report_html"
+        echo "<td>"$encrypted"</td>"  >> "$report_html"
+        echo "<td>$keyCount</td>" >> "$report_html"
+        echo "</tr>"  >> "$report_html"
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
