+++ tools/apigee-sackmesser/cmd/report/helpers/targetservers.sh
@@ -36,25 +36,27 @@ echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/target-servers".json | while read i; do 
-    name=$(echo "$i" | jq -r '.name')
-    _enabled=$(echo "$i" | jq -r '.isEnabled')
-    host=$(echo "$i" | jq -r '.host')
-    port=$(echo "$i" | jq -r '.port')
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/target-servers".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/target-servers".json | while read i; do 
+        name=$(echo "$i" | jq -r '.name')
+        _enabled=$(echo "$i" | jq -r '.isEnabled')
+        host=$(echo "$i" | jq -r '.host')
+        port=$(echo "$i" | jq -r '.port')
 
-    if [ $_enabled = true ]
-        then
-            isEnabled="✅"
-        else
-            isEnabled="❌"
-    fi
+        if [ $_enabled = true ]
+            then
+                isEnabled="✅"
+            else
+                isEnabled="❌"
+        fi
 
-    echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-    echo "<td>$name</td>"  >> "$report_html"
-    echo "<td>"$host"</td>"  >> "$report_html"
-    echo "<td>$port</td>" >> "$report_html"
-    echo "<td>$isEnabled</td>" >> "$report_html"
-    echo "</tr>"  >> "$report_html"
-done
+        echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+        echo "<td>$name</td>"  >> "$report_html"
+        echo "<td>"$host"</td>"  >> "$report_html"
+        echo "<td>$port</td>" >> "$report_html"
+        echo "<td>$isEnabled</td>" >> "$report_html"
+        echo "</tr>"  >> "$report_html"
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
