+++ tools/apigee-sackmesser/cmd/report/helpers/virtualhosts.sh
@@ -36,18 +36,20 @@ echo "</tr></thead>" >> "$report_html"
 
 echo "<tbody class=\"mdc-data-table__content\">" >> "$report_html"
 
-jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/virtualhosts".json | while read i; do 
-    name=$(echo "$i" | jq -r '.name')
-    hostAliases=$(echo "$i" | jq -r '.hostAliases[]')
-    port=$(echo "$i" | jq -r '.port')
-    useBuiltInFreeTrialCert=$(echo "$i" | jq -r '.useBuiltInFreeTrialCert')
+if [ -f "$export_folder/$organization/config/resources/edge/env/$environment/virtualhosts".json ]; then
+    jq -c '.[]' "$export_folder/$organization/config/resources/edge/env/$environment/virtualhosts".json | while read i; do 
+        name=$(echo "$i" | jq -r '.name')
+        hostAliases=$(echo "$i" | jq -r '.hostAliases[]')
+        port=$(echo "$i" | jq -r '.port')
+        useBuiltInFreeTrialCert=$(echo "$i" | jq -r '.useBuiltInFreeTrialCert')
 
-    echo "<tr class=\"$highlightclass\">"  >> "$report_html"
-    echo "<td>$name</td>"  >> "$report_html"
-    echo "<td>$hostAliases</td>" >> "$report_html"
-    echo "<td>$port</td>" >> "$report_html"
-    echo "<td>$useBuiltInFreeTrialCert</td>" >> "$report_html"
-    echo "</tr>"  >> "$report_html"
-done
+        echo "<tr class=\"$highlightclass\">"  >> "$report_html"
+        echo "<td>$name</td>"  >> "$report_html"
+        echo "<td>$hostAliases</td>" >> "$report_html"
+        echo "<td>$port</td>" >> "$report_html"
+        echo "<td>$useBuiltInFreeTrialCert</td>" >> "$report_html"
+        echo "</tr>"  >> "$report_html"
+    done
+fi
 
 echo "</tbody></table></div>" >> "$report_html"
